refactor: remove unused code across backend and frontend
- Replace unused variables with underscore in os.walk loops (storage_service, transcription_manager)

- Remove unused client assignment in feed parser test

- Remove unused variables in test_podcast_bulk_delete

- Remove unused mock patch context in test_transcription

- Remove unused subscription variable in duplicate detection test

- Simplify assertion in password reset test

- Remove unnecessary type cast in home_page

- Remove fallback nil coalescing in podcast_discover_provider

- Remove unused borderColor variable in country_selector_dropdown

- Remove unused _buildLoadingCover method in floating_player_widget

- Remove unused theme variable in podcast_bottom_player_widget

- Remove unused _isMenuOpen state and setState calls in profile_page

- Remove unused app_colors import in update_dialog

- Reorganize imports in conversation_service and tasks/runtime

- Add missing newlines at EOF in test files

+++ frontend/lib/features/home/presentation/pages/home_page.dart
@@ -35,7 +35,7 @@ class _HomePageState extends ConsumerState<HomePage> {
   @override
   void initState() {
     super.initState();
-    _currentIndex = (widget.initialTab ?? 0).clamp(0, _tabCount - 1) as int;
+    _currentIndex = (widget.initialTab ?? 0).clamp(0, _tabCount - 1);
     _visitedTabs.add(_currentIndex);
     WidgetsBinding.instance.addPostFrameCallback((_) {
       if (!mounted) return;
